+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # Change Log
 
+## 0.30.5
+
+- Fix .overlap(SET/BOOLEAN)
+
 ## 0.29.3
 
 - Fix Plywood Set constructor function does not handle null/"null" and falsy values properly
 
 ## 0.29.2
 
-- Add support for converting inExpression to SQL 
+- Add support for converting inExpression to SQL
 
 ## 0.29.1